possibly done refactoring
